requirements.txt
python-dotenv
pymupdf
pdfplumber
pytesseract
Pillow
tqdm


requests>=2.31.0
beautifulsoup4>=4.12.0